refactor(mesh-v2): extract polling-client mixin
issue #566 の責務分割 3/N。Polling モード (WebSocket フォールバック) の
2 秒周期データ取得を `polling-client.js` に切り出し。

切り出し対象:
- startPolling
- stopPolling
- pollEvents (POLL_GROUP_DATA query を発行する本体)

mesh-service.js: 1278 → 1182 行 (-96)
polling-client.js: 106 行 (新規)
